Fix the curl redirects of the buildbot, adding the -L switch. So that it actually follows the location headers and downloads the `upload-to-bitbucket.sh` script from my restored git repo, which even has the original `default` hg branch.
Add an additional --fail parameter in full form to at least catch 40X or 50X server errors and output them to the log.
We won't catch lack-of-redirect bugs like the 302 issue with a lacking -L, though.

Hopefully all these changes should make the uploader work again, before reinstating the repository it returned a 404 and ran the result as a Bash script,
and before doing this it saved the empty 302 Found response from the server.