Rewrite game_get_skill_modifier_for_troop from scratch
Now we separate the checks for the player (Inventory items) from those
of more 'general' heroes, where we have to take into account equipment
and injuries.

This assumes that non-heroes have no special items or wounds.

Further improvements are: check inventory for item and bail as soon as
possible, check for wounds and gate all the wounds behind a check of is
wounded.