Add mipmaps to most of the mountain backdrops to avoid jitter, there are probably other textures with the same issues. Also, the mordormountain.dds file was saved with the alpha-cutout-is-black trick, but renders as transparent in our case, so those pixels show up as holes. If HDR is enabled the sky shows through and causes fireflies and twinkles.
Now with mipmapping the holes get averaged out because the texture is far in the distance, showing MIP 2 or 3 and this doesn't seem that big of an issue. Ideally I'd remove that one bit alpha altogether, but I don't want to mess with the encoding and degrade the texture any further. We don't have the original file.

NVIDIA Texture Exporter ideally should support something like this. Some checkbox to ignore the alpha channel.