Lords will now give some troops as a friendship reward.
When turning in a quest lords will decided (based on relation) if they want to give the player an extra reward in the form of elite troops. Lords with a retainer assigned will give them (if they have any), otherwise they will give the highest level subfaction appropriate troop in their party.