Get rid of the `sf_stream_from_hd` flag in `module_sounds.py` to try to minimize the (less polished) Linux and macOS versions of the game leaking opened file handles for .wav files and reaching the operating system's limit (which tends to be 1024) after a while, causing a `EMFILE` (Too many open files) error on every `fopen()` call. Meaning that from that point onwards the game can no longer save or load, creating black screens and missing sounds and other problems, especially after changing gameplay modes.
Most of this has been found by `null-painter-error#9774` on Discord, I would not have been able to find the cause otherwise, they made it very easy thanks to their amazing Linux knowledge, detective work and responsiveness. Major props!

I may or may not have introduced this flag (and problem) in 2014, in the middle of the Warband port, while trying to reduce the memory usage:
https://github.com/tldmod/tldmod/commit/efb4851e0e3394b7410767fe9ff3b07996d41182 🐧